Fix a false negative with reference in conditional in UselessAssignment
This fixes #475.
